Nav bar and our services
I've added the remaining icons to the nav bar, as well as tweaked the spacing and fonts. I'm now working on the spacing for the 'our services' section and I'm slowly losing my mind trying to get the small inner 'buttons' to align along the bottom of the bigger buttons.