Deadspace and minor tweaks
I've fixed the issue where the icons would load in before the CSS and create a tonne of deadspace at the bottom of the webapge (I just had to change the order of the elements in the <head> tag so that the CSS would load in first)
I tweaked a bit more of the layout and styling to closer match the target site.
I have failed to get the nav-bar first child to have a wider triangle on hover at the md breakpoint.