Submitting after tweaks
I've actioned the feedback I received from the other Scions as much as the time I had left allowed. The News section is pretty rough but the case studies pops up are even worse.

Added more to the README, it should now contain the majority of issues I still have left to sort.